Fix image deletion logic and prevent initial update in question edit

+++ frontend/src/main/frontend/src/components/custom/QuizQuestionExtension/AttachedImageExtensions.tsx
@@ -46,7 +46,9 @@ export const AttachedImageEdit: React.FC<EditProps> = ({
       if (onExtensionChange) {
         onExtensionChange({ ...extension, imageUrl: newImageUrl.fileURL });
       }
-      await deleteFile(oldImageUrl)
+      if (oldImageUrl !== "") {
+        await deleteFile(oldImageUrl)
+      }
     }
     setLoading(false);
   };
